move modal sources: host-copy-network
